refactor: split up settings and contactsection
